Improve reliability of remote matching.
Instead of looking for the literal URL, decompose both the URL that is
being searched for, and each of the remote URLs, and compare the tuple
instead.

This allows matching git remotes when looking for https URLs, and work even
when the user has configured something like:

    url.git@github.com:.pushinsteadof=http://github.com/
    url.git@github.com:.pushinsteadof=https://github.com/
    url.git@github.com:.pushinsteadof=git://github.com/

Also, apply a lower() call to the URL when decomposing it. While I don't
know Pagure that well, GitHub is case insensitive on both the username and
repository names.