Use an explicit attrs-based object as repository identifier.
The 4-tuple is unwieldy to access, and only works for trivial
comparisons. There are a few more places where an actual repository
identifier needs to be used, that supports the subset of case-insensitive
comparisons needed.

Note that this changes a bit the way the repositories are identified. In
particular, instead of lower-casing the whole URL before splitting it, the
URL's case is preserved, but when comparing identifiers, the comparison is
case-insensitive.

This allows preserving the upper/lower case distinction which matters in
_some_ of the GitHub APIs.