Use the user (or org) in the fork response when creating the pull request.
GitHub usernames are case-insensitive in _most_ contexts, but not all.

For example, one might login as `username`, while the "canonical" name is
`UserName`. But the pull requests need to be created with the correct
canonical case for the username.

This re-parses the repository as forked, and preserves the case of the user
(or org) name provided there.

See Issue #100 for details.