Handle repositories with multiple pull request templates.
According to [the
documentation](https://help.github.com/en/github/building-a-strong-community/creating-a-pull-request-template-for-your-repository),
PULL_REQUEST_TEMPLATE can be a directory with multiple `.md` or `.txt`
files. Unfortunately this does not allow selecting which one. And it's
actually incorrect because as far as I can tell, the full path is matched
case-insensitive.

But it looks like it's hard to have a case-insensitive glob in Python, so
for now accept it as is.